+++ x/evm/keeper/statedb_test.go
@@ -656,7 +656,7 @@ func (suite *KeeperTestSuite) TestAddLog() {
 	msg2.From = addr.Hex()
 
 	ethTx3Params := &types.EvmTxArgs{
-		ChainID:   big.NewInt(1),
+		ChainID:   big.NewInt(9000),
 		Nonce:     0,
 		To:        &suite.address,
 		Amount:    big.NewInt(1),